Move functions related to newrelic segments
